fix: fix sft resolve. The resolve handler was not specified for the sft This is the first pass at having a resolve for _call methods it returns a Vec<U128> of amounts per tokenId.
This was done instead of a enumy type of U128/boolean
to support the fungible tokens that might return an
amount as well as the non fungible tokens return 1.

For clarity, if a vector of 0's is returned, then
no additional amounts were returned to user, if
there are non-zero entries, then those entries corresponding
to the token id request order, means that amount was returned.

For NFTs this amount is 1 or 0.