fix: refactor macros to make for an optional minting
The idea here is that macros and core_impl should make
minting easy, but optional for people using the lib.

The lib exports to macros one that auto fills everyting but
minting and one with minting baked in. This will allow
developers to get the core benefit of minting without
necessarily needing to use the core impl's version