feat: adds storage management to semi fungible tokens. This was necessary because the storage management prior to this relied on the minter kind of paying and was not quite working
A new storage management specifically for semi fungible tokens was
required because the sft has a shape that similar to ft and as such
we can have a similar storage management.

Additionally sft semantics are about bulk transfers, and this
will support allocating storage for many sft's at once so for the
use case of transfering silver,gold, and say copper to a account
you can register the user for these tokens all at once as opposed
to 3 txs.

This impl also allows for blanket specification of ft/nft tokens
within the request without breaking functionality, this means if
a consumer building a dapp on top has some index of owners to
tokens, they can quickly unregister the user from groups without
needing to understand what they are unregistering.

For tokens that have a nft type of semantic where there is only
one, the refunds are 0 as the cost is created when minting and
it would be up to the application to refund a nft if it were ever
burnt.